+++ src/esm/components/navigation/header-nav/components/quick-access-link.js
@@ -9,7 +9,7 @@ import { pxToRem } from '../../../../helpers/utils/typography';
 var StyledQuickNav = styled.a.withConfig({
   displayName: "quick-access-link__StyledQuickNav",
   componentId: "sc-88k1zh-0"
-})(["position:absolute;display:inline-flex;align-items:center;left:-100%;top:0;z-index:", ";padding:0 ", ";background-color:", ";border-right:", " solid ", ";", " color:", ";line-height:1;font-size:", ";text-decoration:none;transition:all .2s ease;transition-delay:0,0;opacity:0;&:focus,&:active{left:0;opacity:1;transition-delay:0,.2s;outline:", " auto -webkit-focus-ring-color;}"], function (_ref) {
+})(["position:absolute;display:inline-flex;align-items:center;left:-100%;top:0;z-index:", ";padding:0 ", ";background-color:", ";border-right:", " solid ", ";", " color:", ";line-height:1;font-size:", ";text-decoration:none;transition:all 0.2s ease;transition-delay:0,0;opacity:0;&:focus,&:active{left:0;opacity:1;transition-delay:0,0.2s;outline:", " auto -webkit-focus-ring-color;}"], function (_ref) {
   var zIndex = _ref.zIndex;
   return zIndex;
 }, pxToRem(30), COLORS.background1, pxToRem(1), COLORS.line1, TYPOGRAPHY.fontStyles.regular, COLORS.font1, pxToRem(16), pxToRem(5));